apps/aqi-bridge/requirements.txt
bleak==0.21.1
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.3
websockets==12.0
httpx==0.27.0
python-json-logger==2.0.7
pytest-asyncio==0.23.5
hypothesis==6.98.15
psutil==5.9.8
pyserial==3.5
